Created a new type of AI
The new AI is meant to be much more competative.
New AI is super focused on nodes and launches full
product line-ups each year. AI also cuts prices, if it sees it's
self not competative enough.

AI makes bigger chips on mature nodes to compete and to earn monay
for new nodes.